Rename service file
